refactor(core/exported): moved packet interfaces to packet.go
